refactor: improve consistency and remove unnecessary abstractions
- Fix circular dependency: register.ts now imports directly from
  @agent/index/agentRegistry instead of barrel export
- Move vars.ts from frontend to utils/files/varsUtils.ts
  (pure utility, breaks @agent -> @frontend cycle)
- Remove 4 single-line re-export index files:
  - frontend/editor/index.ts
  - frontend/system/index.ts
  - agent/utils/text/index.ts
  - historyView/managers/index.ts
- Update all imports to use direct paths
- Remove unused exports from xmlUtils.ts (outputFormat, filterTagsFromText)
- Fix broken export in frontend/index.ts (referenced deleted rules.ts)
