--- src/frontend/editor/index.ts
@@ -1,1 +1,0 @@
-export * from './activeFileGuards';
