+++ src/frontend/files/index.ts
@@ -1,4 +1,3 @@
 // Barrel export for frontend file utilities
 export { ListableFileType, FileLister, fileLister } from './fileLister';
 export { getFilesInDirectory, getFilesRecursively } from './listing';
-export { setVarFromFile } from './vars';
