--- src/frontend/system/index.ts
@@ -1,1 +1,0 @@
-export * from './commandUtils';
