--- src/historyView/managers/index.ts
@@ -1,1 +1,0 @@
-export { AgentHistoryManager, type AgentHistoryItem } from '@common/history';
