+++ src/utils/files/varsUtils.ts
